requirements_production.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-secret-manager==2.*
requests==2.*
urllib3==2.*
